requirements.txt
pymupdf>=1.23.0
tqdm>=4.66.0
markdown>=3.5

# Optional — required only for format 7 (OCR):
# Also install the Tesseract binary:
#   macOS:  brew install tesseract tesseract-lang
#   Ubuntu: sudo apt install tesseract-ocr tesseract-ocr-ita
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
pytesseract>=0.3.10
pillow>=10.0.0
